refactor(data-structures): modify ChainInfoError and BuildersError to failure::Error
